third_party/blink/web_tests/virtual/deprecated-css-appearance-values-enabled/css-parser/deprecated-css-appearance-values-warning/appearance-warning-cssom-usecounter-expected.txt

CONSOLE WARNING: The keyword 'inner-spin-button' specified to an 'appearance' property is not standardized. It will be removed in the future.
This is a testharness.js-based test.
[PASS] appearance-warning-cssom-usecounter
Harness: the test ran to completion.
